@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 3.2.0+28
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -28,7 +28,7 @@ dependencies:
   firebase_database: ^7.0.0
   firebase_messaging: ^10.0.0
   fl_chart: ^0.40.2
-  flutter_bloc: ^4.0.1
+  flutter_bloc: ^8.0.1
   flutter_share: ^2.0.0
   flutter_svg: ^0.22.0
   # flutter_webview_plugin: ^0.4.0
@@ -38,7 +38,7 @@ dependencies:
   mantras:
     git:
       url: https://github.com/uttusharma/mantras.git
-  provider: ^4.0.5
+  provider: ^6.0.2
   share: ^2.0.1
   shared_preferences: ^2.0.5
   shimmer: ^2.0.0
