@@ -1,5 +1,5 @@
 class Cred {
-  String username;
-  String password;
+  String? username;
+  String? password;
   Cred({this.username, this.password});
 }
