@@ -7,15 +7,15 @@ import '../../../data/model/hangout/PollModel.dart';
 import '../../../theme/colors.dart';
 
 class PollView extends StatefulWidget {
-  final PollModel poll;
-  final String user;
-  final String postId;
+  final PollModel? poll;
+  final String? user;
+  final String? postId;
   final isCreate;
   const PollView(
-      {Key key,
-      @required this.poll,
+      {Key? key,
+      required this.poll,
       this.postId,
-      @required this.user,
+      required this.user,
       this.isCreate = false})
       : super(key: key);
   @override
@@ -23,7 +23,7 @@ class PollView extends StatefulWidget {
 }
 
 class _PollViewState extends State<PollView> {
-  PollModel poll;
+  PollModel? poll;
 
   final firebaseRef =
       FirebaseDatabase.instance.reference().child('hangout/pings');
@@ -31,7 +31,8 @@ class _PollViewState extends State<PollView> {
   void initState() {
     super.initState();
     poll = widget.poll;
-    if (poll.userWhoVoted == null) poll.userWhoVoted = Map<String, dynamic>();
+    if (poll!.userWhoVoted == null)
+      poll!.userWhoVoted = Map<String?, dynamic>();
   }
 
   @override
@@ -43,9 +44,10 @@ class _PollViewState extends State<PollView> {
         children: [
           // This cannot be less than 2, else will throw an exception
           Polls.options(
-              title: poll.optionLabel.first, value: poll.numberOfVotes.first),
+              title: poll!.optionLabel.first,
+              value: poll!.numberOfVotes.first!),
           Polls.options(
-              title: poll.optionLabel[1], value: poll.numberOfVotes[1]),
+              title: poll!.optionLabel[1], value: poll!.numberOfVotes[1]!),
         ],
         question: Text(''),
         onVoteBackgroundColor: Colors.blue,
@@ -57,18 +59,19 @@ class _PollViewState extends State<PollView> {
         children: [
           // This cannot be less than 2, else will throw an exception
           Polls.options(
-              title: poll.optionLabel.first, value: poll.numberOfVotes.first),
+              title: poll!.optionLabel.first,
+              value: poll!.numberOfVotes.first!),
           Polls.options(
-              title: poll.optionLabel[1], value: poll.numberOfVotes[1]),
+              title: poll!.optionLabel[1], value: poll!.numberOfVotes[1]!),
         ],
         question: Text(
           '',
           style: TextStyle(fontSize: 1),
         ),
         currentUser: widget.user,
-        creatorID: poll.creator,
-        voteData: poll.userWhoVoted,
-        userChoice: poll.userWhoVoted[widget.user],
+        creatorID: poll!.creator,
+        voteData: poll!.userWhoVoted,
+        userChoice: poll!.userWhoVoted![widget.user],
         onVoteBackgroundColor: Colors.blue,
         leadingBackgroundColor: Colors.blue,
         outlineColor: kBlue,
@@ -76,19 +79,21 @@ class _PollViewState extends State<PollView> {
         onVote: (choice) {
           print(choice);
           setState(() {
-            poll.userWhoVoted[widget.user] = choice;
+            poll!.userWhoVoted![widget.user] = choice;
           });
           if (choice == 1) {
             setState(() {
-              poll.numberOfVotes.first += 1.0;
+              var res = poll!.numberOfVotes.first;
+              if (res != null) res += 1.0;
             });
           }
           if (choice == 2) {
             setState(() {
-              poll.numberOfVotes[1] += 1.0;
+              var res = poll!.numberOfVotes[1];
+              if (res != null) res += 1.0;
             });
           }
-          firebaseRef.child(widget.postId).update({'pollData': poll.toMap()});
+          firebaseRef.child(widget.postId!).update({'pollData': poll!.toMap()});
         },
       );
   }
