@@ -2,15 +2,15 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
 class SizeConfig {
-  static double screenWidth;
-  static double screenHeight;
+  static double? screenWidth;
+  static late double screenHeight;
   static double _blockWidth = 0;
   static double _blockHeight = 0;
 
-  static double textMultiplier;
-  static double imageSizeMultiplier;
-  static double heightMultiplier;
-  static double widthMultiplier;
+  static late double textMultiplier;
+  static double? imageSizeMultiplier;
+  static late double heightMultiplier;
+  static late double widthMultiplier;
   static bool isPortrait = true;
   static bool isMobilePortrait = false;
 
@@ -19,7 +19,7 @@ class SizeConfig {
       screenWidth = constraints.maxWidth;
       screenHeight = constraints.maxHeight;
       isPortrait = true;
-      if (screenWidth < 450) {
+      if (screenWidth! < 450) {
         //      print('mobile portrait is true');
         isMobilePortrait = true;
       }
@@ -30,7 +30,7 @@ class SizeConfig {
       isMobilePortrait = false;
     }
 
-    _blockWidth = screenWidth / 100;
+    _blockWidth = screenWidth! / 100;
     _blockHeight = screenHeight / 100;
 
     textMultiplier = _blockHeight;
