@@ -4,12 +4,12 @@ import '../data/const.dart';
 import '../data/model/User.dart';
 
 class SugarParser {
-  User user(String body, username) {
+  User user(String? body, username) {
     var document = parse(body);
-    var name = document.querySelector("#lblname").text;
+    var name = document.querySelector("#lblname")!.text;
 
     var userImage =
-        document.querySelector("#UserImage").attributes.values.elementAt(1);
+        document.querySelector("#UserImage")!.attributes.values.elementAt(1);
     if (name != null) {
       if (userImage == null) {
         userImage = kDefaultUserImage;
