@@ -6,7 +6,7 @@ class SvgWrapper {
 
   SvgWrapper(this.rawSvg);
 
-  Future<DrawableRoot> generateLogo() async {
+  Future<DrawableRoot?> generateLogo() async {
     assert(rawSvg != null);
     try {
       return await svg.fromSvgString(rawSvg, rawSvg);
